backend/crop-advisory/requirements.txt
flask
flask-cors
tensorflow
pandas
numpy
scikit-learn
joblib
google-cloud-vision